python/pyproject.toml
[project]
name = "mindctrl"
version = "0.0.1"
authors = [
  { name="AK" },
]
description = "python package for mindctrl source"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache",
    "Operating System :: OS Independent",
    "Private :: Do Not Upload" # Currently just for local + cluster packaging. Publish once evals are in
]
dependencies = [
  "dapr",
  "dapr-ext-fastapi",
  "dapr-ext-workflow",
  "mlflow-skinny"
]

[project.scripts]
mindctrl = "mindctrl.cli:cli"

[project.urls]
Homepage = "https://github.com/akshaya-a/mindctrl"
Issues = "https://github.com/akshaya-a/mindctrl/issues"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.ruff]
extend-exclude = ["openai_deployment"]
